Regenerate initial migration in dummy app
1909d1b3 changed the limit within the spec's copy of the initial
migration, but not the template. The template migration should be the
canonical one, so if this change was intentional, it should have been
made in the template migration as well. Since this change seems
unrelated, I'm assuming it wasn't intentional, and thus reverting it to
align with the template migration.

The easiest way to ensure these two migrations stay in sync is to only
edit the template migration, and then update the dummy one by running
`rails generate shortener --force` from within the dummy app.