Move retry block outside of ActiveRecord's persistence life cycle
Previously we were catching `ActiveRecord::RecordNotUnique` error in the
`around_save` callback of `ShortenedUrl`. This means the retry was called
inside transaction. It worked fine for MySQL and Sqlite but
Postgres transactions needs to be ROLLBACK'ed on errors.

This is a breaking change if someone expects that
`Shortener::ShortenedUrl::create!` will handle the collisions on `unique_key`.
The collisions will only be handled by `Shortener::ShortenedUrl::generate!`.